chore: collapsed entities in same file
