+++ src/main/scala/scatan/model/DevelopmentCard.scala
@@ -2,4 +2,11 @@ package scatan.model
 
 import scatan.mvc.lib.Model
 
+enum DevelopmentType:
+  case Knight
+  case RoadBuilding
+  case YearOfPlenty
+  case Monopoly
+  case VictoryPoint
+
 final case class DevelopmentCard(developmentType: DevelopmentType)
