--- src/main/scala/scatan/model/DevelopmentType.scala
@@ -1,10 +1,0 @@
-package scatan.model
-
-import scatan.mvc.lib.Model
-
-enum DevelopmentType:
-  case Knight
-  case RoadBuilding
-  case YearOfPlenty
-  case Monopoly
-  case VictoryPoint
