--- src/main/scala/scatan/model/ResourceType.scala
@@ -1,10 +1,0 @@
-package scatan.model
-
-import scatan.mvc.lib.Model
-
-enum ResourceType:
-  case Wood
-  case Brick
-  case Sheep
-  case Wheat
-  case Rock
